chahua: 重构 renderer.js（二）—— 抽出 connection 模块
WebSocket 连接 + 重连退避抽到 connection.js：模块持有 socket 与 connected 标志，
重连决策（退避梯度 / NO_RECONNECT_CODES）整个封进去。renderer 经 send /
isConnected 读写连接，经 onConnecting / onMessage / onClose 三个回调接生命周期；
断线时 renderer 侧的状态归位留在 handleDisconnect。

renderer 内 connected 读点统一改走 connection.isConnected()，裸 ws.send(JSON.
stringify(...)) 统一改走 send(...)。renderer.js 1470 → 1422 行。行为零变化，
node --check 通过。
